Move cache populating task to separate module
